Fix: use enzyme `is` for `match` assertion
Using the `match` assertion appears to be deferring to the `is` function, which
takes any valid EnzymeSelector. However, when passing a react component by
class object (not string) the matcher fails. This is because the `is` matcher
for the ShallowTestWrapper was rendering the full HTML so it could also handle
pseudo-classes (e.g. ':checked').

Given that EnzymeSelectors don't officially support psuedo-classes AND there
are completely separate matchers for those already, I think it makes sense to
make `match` directly use the default `is` function to be more consistent.

This change didn't break any specs, and I've added some new ones that were
failing before the change.